feat: detach component
