+++ src/pages/index.tsx
@@ -4,18 +4,17 @@ import Link from "next/link";
 import Lottie from "react-lottie-player";
 
 import lottieJson from "public/lotties/scrollDown.json";
-import Button from "@/components/Button";
-import { H1, T1, T2 } from "@/components/Text";
+import { H1, T1 } from "@/components/Text";
 import { DebutGroupStore, SelectedMemberStore } from "@/store/store";
 import Seo from "@/components/Seo";
 import useObserver from "hooks/useObserver";
 import MainFirstSection from "@/components/MainFirstSection";
+import MainSecondSection from "@/components/MainSecondSection";
 
 const Home: NextPage = () => {
   const { setSelectedMembers } = SelectedMemberStore();
   const { setDebutGroup } = DebutGroupStore();
   const [isMainSeen, setIsMainSeen] = useState(true);
-  const [isSecondSectionSeen, setIsSecondSectionSeen] = useState(false);
 
   useEffect(() => {
     localStorage.removeItem("scrollY");
@@ -23,7 +22,6 @@ const Home: NextPage = () => {
     setDebutGroup([], "", "");
   }, []);
 
-  //메인 화면의 io
   const onIntersectMain: IntersectionObserverCallback = ([entry]) => {
     entry.isIntersecting ? setIsMainSeen(true) : setIsMainSeen(false);
   };
@@ -31,18 +29,6 @@ const Home: NextPage = () => {
   const { setTarget: setTargetMain } = useObserver({
     onIntersect: onIntersectMain,
     threshold: 0,
-  });
-
-  //2번째 섹션의 io
-  const onIntersectSecondSection: IntersectionObserverCallback = ([entry]) => {
-    console.log(entry.isIntersecting);
-    // if (entry.isIntersecting)
-    entry.isIntersecting ? setIsSecondSectionSeen(true) : setIsSecondSectionSeen(false);
-  };
-
-  const { setTarget: setTargetSecondSection } = useObserver({
-    onIntersect: onIntersectSecondSection,
-    threshold: [0.1, 0.5],
   });
 
   return (
@@ -72,34 +58,8 @@ const Home: NextPage = () => {
           />
         </div>
       </section>
-
       <MainFirstSection />
-
-      <section className="h-[200vh] w-full relative flex flex-col justify-start items-center">
-        <div className="w-full h-[100vh] sticky top-0 z-10 gap-2 flex flex-col justify-center items-center">
-          <span className="w-full text-4xl text-center text-white">원하는 멤버만 골라,</span>
-          <span className="w-full text-4xl text-center text-white">마음대로 만드는</span>
-          <span className="bg-white px-4 py-2 mx-2 text-5xl font-bold rounded-lg text-PRIMARY">
-            나만의 아이돌
-          </span>
-        </div>
-        <div
-          className="h-screen w-full bg-PRIMARY sticky top-0 left-0 flex flex-col justify-center items-center"
-          ref={setTargetSecondSection}
-        ></div>
-      </section>
-
-      <Link href="/casting">
-        <div
-          className={
-            isSecondSectionSeen ? "animate-fadein transition-all" : "animate-fadeout opacity-0 "
-          }
-        >
-          <Button isWhite={true} isFixed={true}>
-            <T2>걸그룹 만들러가기 🎀</T2>
-          </Button>
-        </div>
-      </Link>
+      <MainSecondSection />
     </div>
   );
 };
